project/requirements.txt
absl-py==0.6.1
appnope==0.1.0
astor==0.7.1
backcall==0.1.0
bleach==3.0.2
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.5
django-rest-framework==0.1.0
djangorestframework==3.9.0
entrypoints==0.2.3
gast==0.2.1
grpcio==1.17.1
h5py==2.9.0
imageio==2.4.1
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.4
numpy==1.15.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.9
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
scikit-image==0.14.1
scipy==1.2.0
Send2Trash==1.5.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2
virtualenv==16.2.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
